+++ parser_library/test/debugging/debugger_test.cpp
@@ -64,7 +64,7 @@ TEST(debugger, stopped_on_entry)
     EXPECT_EQ(std::string_view(sc.item(1).name), "Locals");
     EXPECT_EQ(std::string_view(sc.item(2).name), "Ordinary symbols");
     auto globs = d.variables(sc.item(0).variable_reference);
-    EXPECT_EQ(globs.size(), 10U);
+    EXPECT_EQ(globs.size(), 12U);
     auto locs = d.variables(sc.item(1).variable_reference);
     EXPECT_EQ(locs.size(), 0U);
 
@@ -192,6 +192,8 @@ struct frame_vars
         this->globals["&SYSTEM_ID"];
         this->globals["&SYSVER"];
         this->globals["&SYSASM"];
+        this->globals["&SYSM_SEV"];
+        this->globals["&SYSM_HSEV"];
     }
     std::unordered_map<std::string, test_var_value> globals;
     std::unordered_map<std::string, test_var_value> locals;